+++ tests/PluginValidate/Requirements/RepositoryRequirementsTest.php
@@ -1,0 +1,83 @@
+<?php
+
+/*
+ * This file is part of the Moodle Plugin CI package.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @copyright Copyright (c) 2015 Moodlerooms Inc. (http://www.moodlerooms.com)
+ * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace Moodlerooms\MoodlePluginCI\Tests\PluginValidate;
+
+use Moodlerooms\MoodlePluginCI\PluginValidate\Plugin;
+use Moodlerooms\MoodlePluginCI\PluginValidate\Requirements\RepositoryRequirements;
+use Moodlerooms\MoodlePluginCI\PluginValidate\Requirements\RequirementsResolver;
+
+/**
+ * @copyright Copyright (c) 2015 Moodlerooms Inc. (http://www.moodlerooms.com)
+ * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class RepositoryRequirementsTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var RepositoryRequirements
+     */
+    private $requirements;
+
+    protected function setUp()
+    {
+        $this->requirements = new RepositoryRequirements(new Plugin('repository_upload', 'repository', 'upload', ''), 29);
+    }
+
+    protected function tearDown()
+    {
+        $this->requirements = null;
+    }
+
+    public function testResolveRequirements()
+    {
+        $resolver = new RequirementsResolver();
+
+        $this->assertInstanceOf(
+            'Moodlerooms\MoodlePluginCI\PluginValidate\Requirements\RepositoryRequirements',
+            $resolver->resolveRequirements(new Plugin('', 'repository', '', ''), 29)
+        );
+    }
+
+    public function testGetRequiredFiles()
+    {
+        $files = $this->requirements->getRequiredFiles();
+
+        $this->assertNotEmpty($files);
+        foreach ($files as $file) {
+            $this->assertInternalType('string', $file);
+        }
+    }
+
+    public function testGetRequiredClasses()
+    {
+        $classes = $this->requirements->getRequiredClasses();
+
+        $this->assertNotEmpty($classes);
+        foreach ($classes as $class) {
+            $this->assertInstanceOf('Moodlerooms\MoodlePluginCI\PluginValidate\Finder\FileTokens', $class);
+        }
+    }
+
+    public function testGetRequiredStrings()
+    {
+        $fileToken = $this->requirements->getRequiredStrings();
+        $this->assertInstanceOf('Moodlerooms\MoodlePluginCI\PluginValidate\Finder\FileTokens', $fileToken);
+        $this->assertEquals('lang/en/repository_upload.php', $fileToken->file);
+    }
+
+    public function testGetRequiredCapabilities()
+    {
+        $fileToken = $this->requirements->getRequiredCapabilities();
+        $this->assertInstanceOf('Moodlerooms\MoodlePluginCI\PluginValidate\Finder\FileTokens', $fileToken);
+        $this->assertEquals('db/access.php', $fileToken->file);
+    }
+}